requirements.txt
# Основные библиотеки для работы бота
aiogram==3.13.1  # Асинхронная библиотека для Telegram ботов
pydantic==2.10.0  # Валидация данных (используется aiogram)

# Для работы с GigaChat (AI ассистент)
gigachat==0.1.27  # Клиент для GigaChat API (версия может отличаться, проверьте актуальную)

# Работа с базой данных
sqlite3  # Встроен в Python, но указываем для полноты

# Логирование и обработка ошибок
loguru==0.7.2  # Удобная библиотека для логирования (альтернатива logging)

# Работа с CSV файлами
csv  # Встроен в Python, но указываем для документации

# Дополнительные утилиты
python-dotenv==1.0.1  # Для загрузки переменных окружения из .env файла
aiofiles==24.1.0  # Асинхронная работа с файлами (полезно для загрузки CSV)

# Опционально: для улучшения производительности и кэширования
cachetools==5.5.0  # Кэширование данных (например, FAQ)

# Для типизации и статического анализа кода
typing-extensions==4.12.2  # Расширенная поддержка типов

ultralytics==8.3.15  # YOLOv8 от Ultralytics